Missing text after restore in SVM fix
Ignore the .sco files. The only change for this fix is Main.sc. It fixes an issue where ScummVM was restoring global90 with a value of 2, when it needs to be 3 to keep both text and speech working. I had to put the change in 'find talker' method because scumm wasn't running the 'Restore' methods, and putting it in 'Narrator::Say' didn't display the very first text box of the game.